Allow route middleware in Express 3.
Express 3 routes are often in the format

app.get('/', middlewareOne, middlewareTwo, requestHandler)

This wasn't allowed by the old module definition. Varargs are
only allowed at the end of a parameter list so I've changed it
to a list of functions that can have any arguments. Less type
checking in these route calls as a result, but worth it to
be able to use route middleware.